Refactor screen structure and unify template apply behavior
