+++ app/core/ipc.ts
@@ -64,7 +64,6 @@ export interface MainApi {
   updateTemplate: (input: TemplateUpdateInput) => Promise<AppSnapshot>;
   deleteTemplate: (templateId: Id) => Promise<AppSnapshot>;
   applyTemplateToDeckItem: (templateId: Id, itemId: Id) => Promise<AppSnapshot>;
-  resetDeckItemToTemplate: (itemId: Id) => Promise<AppSnapshot>;
   applyTemplateToOverlay: (templateId: Id, overlayId: Id) => Promise<AppSnapshot>;
   renameLibrary: (id: Id, name: string) => Promise<AppSnapshot>;
   renamePlaylist: (id: Id, name: string) => Promise<AppSnapshot>;
@@ -132,7 +131,6 @@ export const IPC = {
   updateTemplate: 'cast:updateTemplate',
   deleteTemplate: 'cast:deleteTemplate',
   applyTemplateToDeckItem: 'cast:applyTemplateToDeckItem',
-  resetDeckItemToTemplate: 'cast:resetDeckItemToTemplate',
   applyTemplateToOverlay: 'cast:applyTemplateToOverlay',
   renameLibrary: 'cast:renameLibrary',
   renamePlaylist: 'cast:renamePlaylist',
