+++ app/main/preload.ts
@@ -70,8 +70,6 @@ const api = {
   deleteTemplate: (templateId: Id) => ipcRenderer.invoke(IPC.deleteTemplate, templateId),
   applyTemplateToDeckItem: (templateId: Id, itemId: Id) =>
     ipcRenderer.invoke(IPC.applyTemplateToDeckItem, templateId, itemId),
-  resetDeckItemToTemplate: (itemId: Id) =>
-    ipcRenderer.invoke(IPC.resetDeckItemToTemplate, itemId),
   applyTemplateToOverlay: (templateId: Id, overlayId: Id) =>
     ipcRenderer.invoke(IPC.applyTemplateToOverlay, templateId, overlayId),
   renameLibrary: (id: Id, name: string) => ipcRenderer.invoke(IPC.renameLibrary, id, name),
