+++ app/renderer/components/controls/button.tsx
@@ -5,7 +5,7 @@ import { cv } from '@renderer/utils/cv';
 export type ButtonVariant = 'default' | 'take' | 'danger' | 'ghost';
 
 const buttonVariants = cv({
-  base: 'cursor-pointer transition-colors px-1 py-px text-center text-xs leading-tight',
+  base: 'cursor-pointer transition-colors px-2.5 py-1 rounded-sm text-center label-xs',
   variants: {
     variant: {
       default: 'bg-tertiary text-primary hover:border-focus hover:text-primary',
