--- app/renderer/features/inspector/inspector-panel.tsx
@@ -1,19 +1,0 @@
-import { Button } from '../../components/controls/button';
-import { Panel } from '../../components/layout/panel';
-import { InspectorTabsPanel } from './inspector-tabs-panel';
-import { useInspectorPanelPushAction } from './use-inspector-panel-push-action';
-
-export function InspectorPanel() {
-  const { state, handlePushChanges } = useInspectorPanelPushAction();
-
-  return (
-    <Panel as="aside" bordered="left" data-ui-region="inspector-panel">
-      <InspectorTabsPanel className="flex-1" />
-      {state.isVisible && <Panel.Footer className="p-3">
-        <Button onClick={handlePushChanges} disabled={state.isPushingChanges} className='w-full'>
-          {state.isPushingChanges ? 'Pushing…' : state.pushLabel}
-        </Button>
-      </Panel.Footer>}
-    </Panel>
-  );
-}
